+++ clientapp/src/main/java/id/co/mii/clientapp/ClientappApplication.java
@@ -1,13 +1,13 @@
-package id.laporyuk.serverapp;
+package id.co.mii.clientapp;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class ServerappApplication {
+public class ClientappApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(ServerappApplication.class, args);
+		SpringApplication.run(ClientappApplication.class, args);
 	}
 
 }
