+++ serverapp/src/main/java/id/co/mii/serverapp/ServerappApplication.java
@@ -1,13 +1,13 @@
-package id.laporyuk.clientapp;
+package id.co.mii.serverapp;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class ClientappApplication {
+public class ServerappApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(ClientappApplication.class, args);
+		SpringApplication.run(ServerappApplication.class, args);
 	}
 
 }
